predefined double view animation rough working
